requirements.txt
numpy
paddlepaddle_gpu>=2.3
Pillow==8.4.0
opencv-python
wandb
